rename AuthContext to ApiProvider
