+++ client/src/pages/ProtectedRoute.tsx
@@ -1,5 +1,5 @@
 import { Navigate } from 'react-router-dom';
-import { useAuthContext } from '../contexts/AuthContext';
+import { useAuthContext } from '../contexts/AuthProvider';
 
 export default function ProtectedRoute({
   children,
